test: always kill warnet server
Currently, if a test fails warnet server might not be properly killed
and so remain running.

This is annoying for iteration; if the test fails we want everything
gone ready to run again.

Also only try to bring the network down if we first brought it up to
avoid spurious error messages in the logs.