patch addrman bucketing, remove build (#261)
* patch: addrman buckets ipv4 based on /32

* images: unified alpine-based dockerfile

* remove k8s dockerfile

* remove image-building scripts

* remove default Dockerfile

* rename alpine dockerfile

* switch to root user exec for alpine

* update image-builder for alpine

* justfile: add build command using new Dockerfile

* image: rehome image_build under cli

* image: fixup image cli command

* compose: add inline branch-build

* ci: disable k8s builder

This is not working currently, but can be re-enabled in the near future
(TM)

* k8s: add inline branch-build

But disable it for now in CI

* server: lock during build stage to permit inline builds

* cli: fix network start command return type

* ci: point minikube local registry to outer docker registry in CI

* images: use new registry with unified image

This image is based on alpine linux, and will work on k8s and compose
backends, as well as forming the base of custom builds.

One image to rule them all, and in the darkness bind them. Or something.