add battalions chart
add chart for deploying a battalion to warnet.
a battalion is a collection of users in a namespace
with roles assigned to the users that allow them to
manage resources in their namespace only. after
the battalion namespace is created, the bitcoincore
chart can be used to deploy tanks to the battalion.