add CVE images with signet (based on #498) (#539)
* add CVE images with signet (based on #498)

add base image for building old versions with CVEs. these images are
pulled from a branch where for each version signet has been packported
and the patches applied.

the image also supports building from a local source, which is useful
when adding new images.

Co-authored-by: Will Clark <6606587+willcl-ark@users.noreply.github.com>

* remove acceptnonstdtx from baseConfig

this option is not allowed on signet. could probably figure out a fix
to allow it but doesnt seem worth it at this time and definitely not something
we want in the base config, anyways.

* fix 16.1 p2sh activation

* remove unused tor section

simplify entrypoint.sh by removing the tor dead code.
we can add this in later if we revisit tor, but its likely
not that we are using helm we will take a different approach
for configuring this, anyways.

* update tests to work with CVE images

Co-authored-by: Matthew Zipkin <2084648+pinheadmz@users.noreply.github.com>

---------

Co-authored-by: willcl-ark <will@256k1.dev>
Co-authored-by: Will Clark <6606587+willcl-ark@users.noreply.github.com>
Co-authored-by: Matthew Zipkin <2084648+pinheadmz@users.noreply.github.com>