users: make the .kube directory creation better
I didn't like the way I originally checked for the directory, then created it -- all the while using
hard-coded values. This one-liner will attempt to make the directory based on KUBECONFIG and if it
exists, then it won't complain. However, if it does not exist, it will create it. This also should
solve the case of the user having a bespoke KUBECONFIG directory in which they deleted their
"config" file. Now, we will use their bespoke path and create the config file in their preferred location.