control: remove `default` namespace from `logs`
A blank namespace is better than using "default". This is because we can figure out the proper
namespace to query against within the logic of the `_logs` function or the user can specify the
namespace as they desire.

Also, Service Account users have no reason to query the `default` namespace.