control: remove the immediate namespace assignment
There is also no reason to eagerly assign a namespace at the beginning of the `_logs` function
because we can figure out the right namespace later on based on context.

Also, we should remove the namespace from the Exception handling because we may not have a namespace
at that point.