control: update format_pods to include namespace
We need a way to filter out pods when the user specifies a namespace using:
`warnet logs --namespace SOME_NAMESPACE`

This is because an admin user may specify a namespace, and we want to show them only those pods that
belong to the namespace. Performing this action in `format_pods` seems natural.