control: get the pod in `warnet logs`
When the user simply runs `warnet logs` or `warnet logs --namespace NS`, we should handle that logic
separately from when they run `warnet logs POD_NAME`. This is because doing so separates the logic of
wrangling multiple namespaces.

When the user runs `warnet logs` or `warnet logs --namespace NS`, the logic of wrangling the
namespaces is very simple: either grab all the namespaces we can, or grab one specific namespace.

When the user runs `warnet logs SOME_POD`, we find ourselves in a situation where we actually need
to query all the namespaces and the whittle down the results to one namespace which contains the pod
OR let the user know that SOME_POD exists in multiple namespaces -- in which case we ask them to
narrow down their query to a specific namespace.