control: wrangle multiple namespaces
When the user runs `warnet logs SOME_POD --namespace NS`, we can actually just try to grab the pod
from that namespace. This is handled in the "else" section towards the end of this commit.

However, when the user does not specify a namespace (`warnet logs SOME_POD`), then we get to query
all the namepaces available to the user, whittle them down, and give the user what they want. We
also ask the user to specify the namespace if we find their desired pod in more than one namespace.