add missing `return`
I noticed that I did not include a `return` if the user simply searches for a non-existent pod. So,
I added the `return` statement. In order to make a test for this, I realized I would need to parse
the output of `warnet logs` to search for "Traceback (" in order to see if there is some kind of
error. Otherwise, if I simple "expect" a message, the test will pass even though the program errors
out with some kind of stack trace.

That is why I made `expect_without_traceback`. It will catch the missing `return` statement. I will
add expect_without_traceback to my other pexpect statements in the next commit.