use `expect_without_traceback` in test
I have extended my test to use `expect_without_traceback`. This means that I swapped out each
`expect` instance with that function, and in doing so, noticed that I needed to watch for `\r` since
we are using inquirer and click which seemingly terminate lines with `\r` as opposed to `\n`.

I also decided to use a short timeout to give the program a chance to generate a Traceback. If a
Traceback is not found within 2 seconds, I then move on to check if we found our expected string and
return that information as a bool.

I also promoted the bitcoin "slug" string to an instance attribute. I also "demoted" the `sut` from an
instance attribute to simply a variable.